crud post, roles

+++ app/Http/Controllers/Admin/PostController.php
@@ -86,6 +86,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('author', $post);
+
         $categories = Category::pluck('name', 'id');//genera un array asociativo con los id y los names.
 
         $tags = Tag::all();
@@ -102,6 +104,8 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+        $this->authorize('author', $post);
+
         $post->update($request->all());
 
         if ($request->file('file')) {
@@ -135,8 +139,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('author', $post);
+
         $post->delete();
 
-        return redirect()->route('admin.posts.index', $post)->with('info', 'El post se eliminó con éxito');
+        return redirect()->route('admin.posts.index')->with('info', 'El post se eliminó con éxito');
     }
 }
